chore: cleaner merge of util.js
